Fix writing null to file on failed save
In the case we failed to convert the model to an xml representation,
we used to write the null to the file. This is particularly bad in
the event an file is overwritten, as all progress will be lost and
failure was silent. While failure is no longer fail dangerous, the
user should still be notified that the saving failed - perhaps by a
simple popup.